[wnsxor1993#28]feat: NetworkService 디렉토리 구성 변경
