+++ iOS/IssueTracker/IssueTracker/Network/NetworkError.swift
@@ -1,0 +1,14 @@
+//
+//  NetworkError.swift
+//  IssueTracker
+//
+//  Created by juntaek.oh on 2022/06/16.
+//
+
+import Foundation
+
+enum NetworkError: Error {
+    case transportError(Error)
+    case serverError(statusCode: Int)
+    case noData
+}
