+++ iOS/IssueTracker/IssueTracker/Network/Service/Protocol/NetworkServiceable.swift
@@ -1,0 +1,13 @@
+//
+//  NetworkServiceable.swift
+//  IssueTracker
+//
+//  Created by juntaek.oh on 2022/06/16.
+//
+
+import Foundation
+
+protocol NetworkServiceable {
+    typealias CompletionHandler = (Result<Data, NetworkError>) -> Void
+    static func request(endPoint: EndPoint, completion: @escaping CompletionHandler)
+}
